requirements.txt
asgiref==3.11.1
channels>=4.0
channels-redis>=4.0
daphne>=4.0
Django==6.0.2
django-cors-headers==4.9.0
django-jazzmin==3.0.2
django-rest-framework==0.1.0
djangorestframework==3.16.1
pillow==12.1.1
sqlparse==0.5.5
tzdata==2025.3
requests>=2.28.0
pycryptodome>=3.18.0
openpyxl>=3.1.0
pymysql>=1.1.0
pdfplumber>=0.10.0
